pyproject.toml
[project]
name = "repl-nix-workspace"
version = "0.1.0"
description = "Add your description here"
requires-python = ">=3.11"
dependencies = [
    "alembic>=1.14.1",
    "pillow>=11.1.0",
    "psycopg2-binary>=2.9.10",
    "python-telegram-bot==20.7",
    "requests>=2.32.3",
    "sqlalchemy>=2.0.38",
    "telegram>=0.0.1",
    "trafilatura>=2.0.0",
]
